l3omda/requirements.txt
flask
flask-sqlalchemy
flask-security-too
psycopg2-binary
gunicorn
email-validator